+++ football-tickets-app/src/main/java/com/ultras/footbalticketsapp/service/TicketServiceImpl.java
@@ -1,7 +1,11 @@
 package com.ultras.footbalticketsapp.service;
 
+import com.ultras.footbalticketsapp.controller.ticket.BuyTicketRequest;
+import com.ultras.footbalticketsapp.controller.ticket.TicketResponse;
+import com.ultras.footbalticketsapp.mapper.TicketMapper;
 import com.ultras.footbalticketsapp.model.Ticket;
 import com.ultras.footbalticketsapp.repository.TicketRepository;
+import com.ultras.footbalticketsapp.serviceInterface.MatchService;
 import com.ultras.footbalticketsapp.serviceInterface.TicketService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,20 +16,30 @@ import java.util.List;
 public class TicketServiceImpl implements TicketService {
 
     private final TicketRepository ticketRepository;
+    private MatchService matchService;
+    private final TicketMapper ticketMapper;
 
     @Autowired
-    public TicketServiceImpl(TicketRepository ticketRepository) {
+    public TicketServiceImpl(TicketRepository ticketRepository, MatchService matchService, TicketMapper ticketMapper) {
         this.ticketRepository = ticketRepository;
+        this.matchService = matchService;
+        this.ticketMapper = ticketMapper;
     }
 
     @Override
-    public Ticket saveTicket(Ticket ticket) {
-        return ticketRepository.save(ticket);
+    public TicketResponse buyTicket(BuyTicketRequest buyTicketRequest) {
+        Ticket ticket = ticketMapper.buyTicketRequestToTicket(buyTicketRequest);
+        if(matchService.TicketBought(ticket.getMatch().getId())){
+            ticketRepository.save(ticket);
+            return ticketMapper.ticketToTicketResponse(ticket);
+        }
+        return null;
     }
 
     @Override
-    public Ticket getTicketById(int id) {
-        return ticketRepository.findById(id).orElse(null);
+    public TicketResponse getTicketById(int id) {
+        Ticket ticket = ticketRepository.findById(id).orElse(null);
+        return ticketMapper.ticketToTicketResponse(ticket);
     }
 
     /*@Override
@@ -34,18 +48,32 @@ public class TicketServiceImpl implements TicketService {
     }*/
 
     @Override
-    public List<Ticket> getTicketsByUserId(int userId) {
-        return ticketRepository.findAllByUserId(userId);
+    public List<TicketResponse> getTicketsByUserId(int userId) {
+        return ticketMapper.ticketsToTicketsResponse(ticketRepository.findAllByUserId(userId));
     }
 
+    //TODO remove because ticket data should not be changed
     @Override
-    public Ticket updateTicket(Ticket ticket) {
-        return null;
+    public TicketResponse updateTicket(TicketResponse ticket) {
+        Ticket updatedTicket = ticketMapper.ticketResponseToTicket(ticket);
+        Ticket ticketToUpdate = ticketRepository.findById(ticket.getId()).orElse(null);
+        if(ticketToUpdate == null){
+            throw new IllegalStateException("Ticket not found");
+        }
+        ticketToUpdate.setMatch(updatedTicket.getMatch());
+        ticketToUpdate.setBuyer(updatedTicket.getBuyer());
+        ticketToUpdate.setPrice(updatedTicket.getPrice());
+        ticketRepository.save(ticketToUpdate);
+        return ticketMapper.ticketToTicketResponse(ticketToUpdate);
     }
 
+    //TODO should be removed because data of every bought ticket should be kept
     @Override
     public void deleteTicket(int id) {
         Ticket ticket = ticketRepository.findById(id).orElse(null);
+        if(ticket == null){
+            throw new IllegalStateException("Ticket not found");
+        }
         ticketRepository.delete(ticket);
     }
 }
